requirements.txt
Django==1.9.6
djangorestframework==3.3.3
tweepy==3.5.0
pytz==2016.6.1
